split the parthenon hdf5 header to reduce LOC included whenever we include params
