+++ src/parthenon_arrays.hpp
@@ -43,12 +43,12 @@ using ParArrayND = ParArrayGeneric<device_view_t<T, Layout>, State>;
 template <typename T, typename State = empty_state_t, typename Layout = LayoutWrapper>
 using ParArrayHost = ParArrayGeneric<host_view_t<T, Layout>, State>;
 
-// #define PARTHENON_ARRAY_DECL(T)                                                          \
-//   extern template class ParArrayGeneric<device_view_t<T, LayoutWrapper>, empty_state_t>
-// 
-// PARTHENON_ARRAY_DECL(Real);
-// 
-// #undef PARTHENON_ARRAY_DECL
+#define PARTHENON_ARRAY_DECL(T)                                                          \
+  extern template class ParArrayGeneric<device_view_t<T, LayoutWrapper>, empty_state_t>
+
+PARTHENON_ARRAY_DECL(Real);
+
+#undef PARTHENON_ARRAY_DECL
 
 } // namespace parthenon
